move error conversions to client
